+++ crates/core/src/string/stack_str.rs
@@ -335,8 +335,8 @@ impl Serialize for StackStr {
 
 impl<'de> Deserialize<'de> for StackStr {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        let s = <&str>::deserialize(deserializer)?;
-        Self::new_checked(s).map_err(serde::de::Error::custom)
+        let s: std::borrow::Cow<'de, str> = Deserialize::deserialize(deserializer)?;
+        Self::new_checked(s.as_ref()).map_err(serde::de::Error::custom)
     }
 }